Move all food related classes to a food package
