+++ src/controller/FilesController.java
@@ -3,20 +3,18 @@ package controller;
 import controller.ViewManager.ViewManagerException;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
-import model.Consumable;
+import model.food.Consumable;
 import model.DatabaseModel;
 import model.User;
 import model.transaction.Transaction;
 
 import java.io.IOException;
-import java.util.ArrayList;
 
 public class FilesController extends Controller
 {
