+++ src/model/food/Col2.java
@@ -1,4 +1,4 @@
-package model;
+package model.food;
 
 public class Col2
 {
