+++ src/model/food/Consumable.java
@@ -1,4 +1,4 @@
-package model;
+package model.food;
 
 import java.util.*;
