+++ src/model/food/ConsumableQuantityPair.java
@@ -1,4 +1,6 @@
-package model;
+package model.food;
+
+import model.food.Consumable;
 
 public class ConsumableQuantityPair
 {
