+++ src/model/food/Meal.java
@@ -1,4 +1,4 @@
-package model;
+package model.food;
 
 import java.util.*;
