+++ src/model/food/RawItem.java
@@ -1,4 +1,4 @@
-package model;
+package model.food;
 
 /**
  * The representation of an item/ingredient in stock in the inventory.
